+++ lib/src/repo_module_creators/module_directory/repo_module_impl_directory_creator.dart
@@ -6,13 +6,15 @@ class RepoModuleImplDirectoryCreator implements RepoModuleIDirectoryCreator {
   final _model = 'model';
   final _views = 'views';
   final _controller = 'controller';
+  final _bloc = "bloc";
   final _components = 'components';
   final _modules = 'modules';
   final _repository = 'repository';
   final _state = 'state';
 
   final String moduleName;
-  RepoModuleImplDirectoryCreator(this.moduleName);
+  final String? modulePattern;
+  RepoModuleImplDirectoryCreator(this.moduleName, this.modulePattern);
 
   late final String basePath;
   late final String projectDirPath;
@@ -33,12 +35,18 @@ class RepoModuleImplDirectoryCreator implements RepoModuleIDirectoryCreator {
 
       print('creating directories...\n');
       //module directory
-      print('creating module directory based on repository pattern...');
+      print(
+          'creating module directory based on ${modulePattern == "1" ? "Bloc pattern" : "Repository pattern"} ...');
       await Directory(absMvcPath).create();
       await Directory('$absMvcPath/$moduleName').create();
+      if (modulePattern == "1") {
+        await Directory('$absMvcPath/$moduleName/$_bloc').create();
+      } else {
+        await Directory('$absMvcPath/$moduleName/$_controller').create();
+        await Directory('$absMvcPath/$moduleName/$_controller/$_state')
+            .create();
+      }
 
-      await Directory('$absMvcPath/$moduleName/$_controller').create();
-      await Directory('$absMvcPath/$moduleName/$_controller/$_state').create();
       await Directory('$absMvcPath/$moduleName/$_model').create();
       await Directory('$absMvcPath/$moduleName/$_repository').create();
       await Directory('$absMvcPath/$moduleName/$_views').create();